Split grpc sim mode into separate file
